perf: GetAllInterfaceTrait is added to implement the interface getAll method
